+++ src/client/components/complexes/project-manage-modal.tsx
@@ -82,7 +82,7 @@ const ProjectManageModal = React.memo<ProjectManageModalProps>(({
   }, [true])
 
   const handleSelectProject = useCallback((value: string) => {
-    data.current.project = value
+    data.current.project = value.split(" ")[0]
     forceUpdate()
   }, [true])
 
@@ -143,17 +143,7 @@ const ProjectManageModal = React.memo<ProjectManageModalProps>(({
     )).filter((project: ProjectInfo) => (
       project.name.includes(data.current.filter) || project.description.includes(data.current.filter)
     )).map((project: ProjectInfo) => (
-      project.name
-    ))
-  )
-
-  const listTitle = () => (
-    data.current.projects.filter((project: ProjectInfo) => (
-      (action === "delete") || (action !== project.status)
-    )).filter((project: ProjectInfo) => (
-      project.name.includes(data.current.filter) || project.description.includes(data.current.filter)
-    )).map((project: ProjectInfo) => (
-      project.description
+      project.name + ((!!project.description) ? ` [ ${ project.description } ]` : "")
     ))
   )
 
@@ -174,7 +164,6 @@ const ProjectManageModal = React.memo<ProjectManageModalProps>(({
           />
           <ListForm
             labels={ listLabel() }
-            titles={ listTitle() }
             onChange={ handleSelectProject }
           />
         </>